+++ imports/plugins/included/shipping-rates/server/index.js
@@ -1,2 +1,3 @@
 import "./i18n";
 import "./methods/rates";
+import "./hooks/hooks";
