+++ imports/plugins/included/shippo/server/index.js
@@ -1,3 +1,4 @@
 import "./methods";
 import "./jobs";
 import "./i18n";
+import "./hooks/hooks";
